Say we're in-sync if the full hash matches
Fixes #68

Previously: If you pointed to a git hash, sometimes `checkout_externals`
said we're out-of-sync when really we're in-sync. This happens when the
abbreviated hash printed by `git branch --verbose --verbose`, (e.g.,
`(HEAD detached at cade91dfe)`) had a different number of characters
than the hash listed in the externals description file. This was a
problem if the full 40-character hash is listed in the externals
description file, or if the number of characters used in the
abbreviation in the externals description file differed from the number
of characters printed by git. Furthermore, running `checkout_externals`
to bring things in sync did not resolve this problem: it continued to
mark the given external as out-of-sync.

With this change, we call things in-sync if the full 40-character hash
is listed in the externals configuration file and we are in detached
head state with that hash checked out.